requirements/local.txt
-r base.txt
django-extensions==1.5.7
django-debug-toolbar==1.4
